Layer calculation and scheduling
We now calculate the number of sections and dynamically determine the
number of layers in each section. The layers are set up to calculate a
set of notes to play, play them, and then exit. This makes figuring out
the actual set of notes to play part of the layer and not global.

This version calculates and schedules the layers but doesn't play
anything in them yet.