Two versions of nopulse
The original (nopulse.ck) always uses a length of 32 for each note, and
plays the notes at the start of the section, giving the dense
clangourous sounds of the version posted on Soundcloud.

The new version (nopulse-async.ck) actually does properly handle the
spreading of the scheduled events, giving a thinner but busier texture.